Restore the default mask and array treatments by netCDF4
After netCDF4 1.4.1, all varaibles will be readed in as numpy masked array and
the mask of a masked array without any missing value will become a scalar.  This
PR will restore the default treatments same as the previous netCDF4 versions.

The restoration method is suggested by the developer. Please check it at:

https://github.com/Unidata/netcdf4-python/blob/master/Changelog