Allows depths to be automatically expanded in the configure language.
With this commit, 'h2' headings now allow the 'depths' keyword which
will accept a float or comma delimited list of floats. If the variable
has a 'depth' dimension, then this keyword will automatically expand
this 'h2' into a list of 'h2', one for each depth specified. The child
nodes are also copied and assigned a 'depth' value to be used in
picking a layer out of the reference dataset.